Rename BasicAuthModule to DefaultAuthModule
It covers not just basic auth but also bearer auth
