+++ yamcs-simulation/etc/privileges.yaml
@@ -7,7 +7,7 @@ enabled: true
 defaultUser: admin
 
 authModule:
-    class: org.yamcs.security.BasicAuthModule
+    class: org.yamcs.security.DefaultAuthModule
     #one of: LdapRealm, YamlRealm
     config:
         realm: org.yamcs.security.YamlRealm
@@ -17,4 +17,3 @@ maxNoSessions: 10
 
 # yaml file, if using YamlRealm
 yamlRealmFilename:  credentials.yaml
-
